Rename `RemoveWildcardImports` to `ForbidWildcardImports` in maven.
