+++ plugin-maven/src/main/java/com/diffplug/spotless/maven/java/ForbidWildcardImports.java
@@ -20,7 +20,7 @@ import com.diffplug.spotless.java.RemoveWildcardImportsStep;
 import com.diffplug.spotless.maven.FormatterStepConfig;
 import com.diffplug.spotless.maven.FormatterStepFactory;
 
-public class RemoveWildcardImports implements FormatterStepFactory {
+public class ForbidWildcardImports implements FormatterStepFactory {
 	@Override
 	public FormatterStep newFormatterStep(FormatterStepConfig config) {
 		return RemoveWildcardImportsStep.create();
